Return database users in our GET /users route.
Notice that we had to add a Models module to define our models.

There is an entire page on how to define our entities in Persistent,
which [can be found
here](https://github.com/yesodweb/persistent/blob/master/docs/Persistent-entity-syntax.md).

Persistent can automatically derive JSON instances (there's a section on
this in the referenced docs).