Protect API with DO NOTHING authentication.
Notice that our `authCheck` function, which is supposed to check the
user's credentials and return an authed user (or not), doesn't do
anything, but we know that everything is wired appropriately.

Notice that, in order to get authentication with Basic Auth working, we
had to:

1. Use the `BasicAuth` combinator in our API.
1. Use `serveWithContext` to create our application.
1. Create a value of type `Context`, used for passing data to
combinators.
1. Create the function that the `BasicAuth` combinator will invoke, with
the credentials, to authenticate a user.